service provider pattern impl: list of order validators autowired correctly
