+++ src/main/java/ch/shukalovi/validator/OrderValidatorExtended.java
@@ -1,0 +1,10 @@
+package ch.shukalovi.validator;
+
+import ch.shukalovi.model.Order;
+
+public class OrderValidatorExtended implements OrderValidator {
+    @Override
+    public boolean isOrderValid(Order order) {
+        return true;
+    }
+}
