+++ src/main/resources/beans/services.xml
@@ -7,11 +7,10 @@
     <bean class="ch.shukalovi.service.OrderService" init-method="init" destroy-method="destroy" autowire="constructor">
 <!--        <constructor-arg ref="orderValidator" />-->
     </bean>
-    <bean id="orderValidator" class="ch.shukalovi.validator.OrderValidator" autowire="constructor">
+    <bean id="orderValidatorBasic" class="ch.shukalovi.validator.OrderValidatorBasic" autowire="constructor">
         <property name="minOrderNo" value="100" />
     </bean>
-    <bean id="orderValidatorCopy" class="ch.shukalovi.validator.OrderValidator" autowire="constructor">
-        <property name="minOrderNo" value="100" />
+    <bean id="orderValidatorExtended" class="ch.shukalovi.validator.OrderValidatorExtended" autowire="constructor">
     </bean>
     <bean id="ntsov" class="ch.shukalovi.validator.NonThreadSafeOrderValidator" scope="prototype"/>
